requirements.txt
tqdm==4.65.0
ultralytics
Pillow
imageio
matplotlib
kornia
natsort
pyyaml
wandb
lpips
open3d==0.16.0
torchmetrics
cyclonedds
pytorch-msssim
plyfile==0.8.1
faiss-cpu==1.8.0,
numpy==1.24.2,
openai==1.33.0,
open_clip_torch==2.26.1,
urllib3==2.2.3,
scipy==1.14.1,
supervision==0.22.0,
httpx==0.27.0,
openai-clip==1.0.1,
clip==0.2.0,
scikit-image    
transformers==4.51.3
git+https://github.com/xinyu1205/recognize-anything.git
git+https://github.com/NVlabs/describe-anything
fairscale
# ace
# pyrender
# trimesh
# pyrealsense2